% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_doc_0.R
\name{html_doc_0}
\alias{html_doc_0}
\title{passthrough for html_document}
\usage{
html_doc_0(...)
}
\arguments{
\item{...}{arguments passed to
\code{rmarkdown::\link[rmarkdown]{html_document}}}
}
\value{
R Markdown output format to pass to
  \code{rmarkdown::\link[rmarkdown]{render}}
}
\description{
The simplest thing we can do is pass all our arguments to
\code{rmarkdown::\link[rmarkdown]{html_document}}
}
\examples{
\dontrun{
  rmarkdown::render("foo.Rmd", html_doc_0())
}
}
\seealso{
\url{http://rmarkdown.rstudio.com/developer_custom_formats.html}
}

