% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_doc_3.R
\name{html_doc_3}
\alias{html_doc_3}
\title{small tweak to html_document}
\usage{
html_doc_3(code_folding = c("show", "hide", "none"), css = NULL, ...)
}
\arguments{
\item{code_folding}{Enable document readers to toggle the
display of R code chunks. We change the default to \code{"hide"}.
See \code{rmarkdown::\link[rmarkdown]{html_document}}}

\item{css}{One or more css files to include in addition
to this format's css.}

\item{...}{other arguments passed to
\code{rmarkdown::\link[rmarkdown]{html_document}}}
}
\value{
R Markdown output format to pass to
  \code{rmarkdown::\link[rmarkdown]{render}}
}
\description{
The next simplest thing we can do is change the defaults for
\code{code_folding} and theme, then pass the rest of our arguments to
\code{rmarkdown::\link[rmarkdown]{html_document}}
}
\examples{
\dontrun{
  rmarkdown::render("foo.Rmd", html_doc_3())
}
}

